chg: dev: travelpoint was split in another file with a new function to list travel points
