+++ modules/bounty.py
@@ -8,7 +8,6 @@ from .mouse_utils import (
     move_mouse,
     mouse_position,
     mouse_click,
-    mouse_scroll,
     mouse_range,
 )
 
@@ -18,7 +17,7 @@ from .game import waitForItOrPass, defaultCase
 from .encounter import selectCardsInHand
 from .campfire import look_at_campfire_completed_tasks
 from .log_board import LogHSMercs
-from .settings import settings_dict, jposition, jthreshold
+from .settings import settings_dict, jthreshold
 from .treasure import chooseTreasure
 from .notification import send_notification
 
@@ -55,7 +54,9 @@ def collect():
         if find_ellement(Button.done.filename, Action.move_and_click):
             break
         if collectAttempts > 10:
-            log.info(f"Attempted to collect treasure {collectAttempts} times, attempting to recover.")
+            log.info(
+                f"Attempted to collect treasure {collectAttempts} times, attempting to recover."
+            )
             break
 
     # move the mouse to avoid a bug where the it is over a card/hero (at the end)
@@ -109,7 +110,9 @@ def nextlvl():
             time.sleep(7)
             while find_ellement(UIElement.visitor.filename, Action.screenshot):
                 if settings_dict["stopatstranger"]:
-                    send_notification({"message": "Stopping after meeting Mysterious Stranger"})
+                    send_notification(
+                        {"message": "Stopping after meeting Mysterious Stranger"}
+                    )
                     log.info("Stopping after meeting Mysterious Stranger")
                     sys.exit()
 
@@ -198,54 +201,6 @@ def searchForEncounter():
     return retour
 
 
-def travelpointSelection():
-    """Choose a Travel Point (Barrens, Felwood, ...)
-    and the mode : Normal or Heroic
-    """
-
-    if find_ellement(UIElement.travelpoint.filename, Action.screenshot):
-
-        move_mouse(windowMP(), windowMP()[2] // 1.5, windowMP()[3] // 2)
-
-        mouse_scroll(jposition["travelpoint.scroll.top"])
-        time.sleep(0.5)
-
-        location = settings_dict["location"]
-        tag = f"travelpoint.{location}.scroll"
-        if location == "Barrens":
-            find_ellement(
-                UIElement.Barrens.filename,
-                Action.move_and_click,
-                jthreshold["travelpoints"],
-            )
-
-        else:
-            try:
-                mouse_scroll(jposition[tag])
-                move_mouse(windowMP(), windowMP()[2] // 3, windowMP()[3] // 2)
-                time.sleep(0.5)
-                find_ellement(
-                    getattr(UIElement, location).filename,
-                    Action.move_and_click,
-                    jthreshold["travelpoints"],
-                )
-            except Exception:
-                log.error(f"Travel Point unknown : {location}")
-
-        move_mouse(windowMP(), windowMP()[2] // 2, windowMP()[3] // 2)
-        time.sleep(0.5)
-
-        if settings_dict["mode"] == "Normal":
-            find_ellement(UIElement.normal.filename, Action.move_and_click)
-        elif settings_dict["mode"] == "Heroic":
-            find_ellement(UIElement.heroic.filename, Action.move_and_click)
-        else:
-            log.error("Settings (for Heroic/Normal) unrecognized.")
-
-    waitForItOrPass(Button.choose_travel, 2)
-    find_ellement(Button.choose_travel.filename, Action.move_and_click)
-
-
 def goToEncounter():
     """
     Start new fight,
@@ -282,8 +237,8 @@ def goToEncounter():
 
             zL = LogHSMercs(settings_dict["zonelog"])
             zL.start()
-            retour = (
-                selectCardsInHand(zL)
+            retour = selectCardsInHand(
+                zL
             )  # Start the battle : the bot choose the cards and fight against the enemy
             zL.stop()
             log.info(f"goToEncounter - retour = {retour}")
